GCC-clang compatibility improvements

+++ src/decoderstrategy.h
@@ -121,7 +121,7 @@ public:
             const int bytesToRead = (bufType_bit_count - _validBits) >> 3;
             _position += bytesToRead;
             _validBits += bytesToRead * 8;
-            ASSERT(_validBits >= bufType_bit_count - 8);
+            ASSERT(static_cast<size_t>(_validBits) >= bufType_bit_count - 8);
             return true;
         }
         return false;
@@ -129,7 +129,7 @@ public:
 
     void MakeValid()
     {
-        ASSERT(_validBits <=bufType_bit_count - 8);
+        ASSERT(static_cast<size_t>(_validBits) <=bufType_bit_count - 8);
 
         if (OptimizedRead())
             return;
